refactor(orchestrator): move `AccountType` declaration to utilties lib
