+++ libs/utilities/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod graphql;
+pub mod users;
 
 use std::{
     env,
