+++ libs/utilities/src/users.rs
@@ -1,0 +1,23 @@
+use std::str::FromStr;
+
+use async_graphql::Enum;
+use strum::Display;
+
+/// The types of accounts a user can create
+#[derive(Enum, Clone, Copy, PartialEq, Eq, Display)]
+pub enum AccountType {
+    Student,
+    Teacher,
+}
+
+impl FromStr for AccountType {
+    type Err = ();
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "Student" | "users::Student" => Ok(Self::Student),
+            "Teacher" | "users::Teacher" => Ok(Self::Teacher),
+            _ => Err(()),
+        }
+    }
+}
